缩减notify微服务的子模块数量 #9230 删除非必要类
