+++ src/backend/ci/core/notify/biz-notify/src/main/kotlin/com/tencent/devops/notify/blueking/service/inner/BlueKingWeworkServiceImpl.kt
@@ -30,8 +30,6 @@ import com.tencent.devops.notify.pojo.WeworkNotifyMediaMessage
 import com.tencent.devops.notify.pojo.WeworkNotifyTextMessage
 import com.tencent.devops.notify.service.WeworkService
 import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty
-import org.springframework.context.annotation.Configuration
 
 class BlueKingWeworkServiceImpl @Autowired constructor() : WeworkService {
