+++ src/backend/ci/core/notify/biz-notify/src/main/kotlin/com/tencent/devops/notify/tencentcloud/config/TencentCloudEmailConfiguration.kt
@@ -26,7 +26,6 @@ class TencentCloudEmailConfiguration {
     ) =
         TencentCloudConfiguration(secretId, secretKey, emailRegion, emailSender)
 
-
     @Bean
     fun emailService(
         @Autowired emailNotifyDao: EmailNotifyDao,
